+++ CHANGELOG.md
@@ -1,2 +1,7 @@
 ## [0.0.1] 
-* 完成地址选择器
+* complete address selector
+
+## [0.0.2] 
+* single column selector
+* Built-in some data
+* 内置了一些数据（性别、学历、学科、星座、生肖、名族）